+++ src/python/pants/bin/pants_exe.py
@@ -17,7 +17,7 @@ from pants.base.build_file_address_mapper import BuildFileAddressMapper
 from pants.base.build_file_parser import BuildFileParser
 from pants.base.build_graph import BuildGraph
 from pants.base.config import Config
-from pants.base.dev_backend_loader import load_build_configuration_from_source
+from pants.base.extension_loader import load_build_configuration_from_source
 from pants.base.rcfile import RcFile
 from pants.base.workunit import WorkUnit
 from pants.commands.command import Command
