+++ tests/python/pants_test/BUILD
@@ -87,7 +87,7 @@ python_tests(
     'src/python/pants/base:build_file_parser',
     'src/python/pants/base:build_graph',
     'src/python/pants/base:config',
-    'src/python/pants/base:dev_backend_loader',
+    'src/python/pants/base:extension_loader',
   ]
 )
